Refactor API server in DefaultScheduler and UninstallScheduler
- Removed unused constants from DefaultScheduler
- No longer check if apiServer is ready in `initializeApiServer()`
